Remove dependency on LFS

+++ src/combat.rs
@@ -77,7 +77,7 @@ pub struct Background;
 
 #[derive(Component)]
 pub struct CombatAgent{
-	pub(crate) q: HashMap<String, HashMap<String, f64>>
+	pub(crate) q: HashMap<String, HashMap<String, isize>>
 }
 
 fn spawn_combat_background(
